refactor(utils): move functions from backend to utils
